requirements.txt
Django>=3.2,<5.1 # Or your specific Django version range
djangorestframework>=3.12,<3.16
django-cors-headers>=3.7,<4.4
djangorestframework-simplejwt>=5.0,<5.4
psycopg2-binary>=2.8,<2.10
gunicorn>=20.0,<22.1
